ElementFinderServiceでWindowHandleフィルタのログレベルをDebugからInformationに変更し、WindowHandleの型をIntPtrからlongに変更しました。また、WindowActionOperationでWindowHandleの取得をlong型に修正し、Program.csにコンソールエラーロギングを一時的に追加しました。

+++ UIAutomationMCP.Models/Results/WindowResults.cs
@@ -7,7 +7,7 @@ namespace UIAutomationMCP.Models.Results
     {
         public string ActionName { get; set; } = string.Empty;
         public string WindowTitle { get; set; } = string.Empty;
-        public IntPtr WindowHandle { get; set; }
+        public long WindowHandle { get; set; }
         public string PreviousState { get; set; } = string.Empty;
         public string CurrentState { get; set; } = string.Empty;
         public bool Completed { get; set; }
